fix translationForm to typescript + refactor
